Add avatar support
