+++ package.json
@@ -23,6 +23,7 @@
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "graphql-tools": "^7.0.1",
+    "graphql-upload": "^11.0.0",
     "jsonwebtoken": "^8.5.1",
     "lodash": "^4.17.20",
     "mailgun-js": "^0.22.0",
