+++ src/modules/users/resolvers/me.js
@@ -1,0 +1,9 @@
+const User = require('../model');
+
+async function me(_, { id }, { authenticate, authorize }) {
+  authenticate();
+  authorize(id);
+  return User.findById(id);
+}
+
+module.exports = me;
